+++ network_io/include/aare/network_io/ZmqWorker.hpp
@@ -5,15 +5,14 @@ namespace aare {
 
 class ZmqWorker {
   public:
-    explicit ZmqWorker(const std::string &ventilator_endpoint, const std::string &sink_endpoint="");
-    Task* pull();
-    int push(const Task *task);
+    explicit ZmqWorker(const std::string &ventilator_endpoint, const std::string &sink_endpoint = "");
+    Task *pull();
+    size_t push(const Task *task);
     ~ZmqWorker();
 
   private:
     ZmqSingleReceiver *m_receiver;
     ZmqSocketSender *m_sender;
 };
-
 
 } // namespace aare
